+++ config/index.js
@@ -61,6 +61,8 @@ module.exports = {
           'form_required_field': 'This field is required',
           'form_validation_failed': 'Validation failed',
           'form_not_a_valid_number': 'Not a valid number',
+          'form_not_a_valid_integer': 'Not a valid integer value',
+          'form_not_a_valid_long': 'Not a valid long value',
           'form_not_a_valid_url': 'Not a valid URL',
           'form_not_a_valid_email': 'Not a valid email',
           'form_not_within_range': 'Value is outside of range',
@@ -68,7 +70,8 @@ module.exports = {
           'form_above_max_value': 'Value is above allowed value',
           'form_bool_true': 'True',
           'form_bool_false': 'False',
-          'form_bool_missing': 'N/A'
+          'form_bool_missing': 'N/A',
+          'form_no_options': 'No options found.'
         }
         res.json(localizedMessages)
       })