+++ src/components/field-types/TypedFieldComponent.vue
@@ -6,13 +6,14 @@
       <input
         :id="field.id"
         v-model="localValue"
-        :type="field.type"
+        :type="inputType"
         :name="field.id"
         class="form-control"
         :class="{ 'is-invalid' : fieldState && (fieldState.$touched || fieldState.$submitted) && fieldState.$invalid}"
         :aria-describedby="field.id + '-description'"
         :required="isRequired"
-        :disabled="field.disabled">
+        :disabled="field.disabled"
+        :step="stepSize">
 
       <small :id="field.id + '-description'" class="form-text text-muted">
         {{ field.description }}
@@ -28,6 +29,12 @@
   import VueForm from 'vue-form'
   import { FormField } from '../../flow.types'
   import FormFieldMessages from '../FormFieldMessages'
+  import debounce from 'debounce'
+
+  const MIN_JAVA_INT = -2147483648
+  const MAX_JAVA_INT = 2147483647
+
+  let debounceTime = 500
 
   export default {
     name: 'TypedFieldComponent',
@@ -55,6 +62,10 @@
       isRequired: {
         type: Boolean,
         default: false
+      },
+      inputDebounceTime: {
+        type: Number,
+        default: debounceTime
       }
     },
     mixins: [VueForm],
@@ -65,21 +76,44 @@
       }
     },
     watch: {
-      localValue (value) {
-        // Emit value changes to the parent (form)
-        this.$emit('input', value)
+      localValue: debounce(function (value) {
+        /*
+        Do not convert NaN field to number to allow for validation to generate warning
+         */
+        if (this.isNumberField(this.field) && !Number.isNaN(value)) {
+          this.$emit('input', Number(value))
+        } else {
+          this.$emit('input', value)
+        }
+
         // Emit value changes to trigger the onValueChange
         // Do not use input event for this to prevent unwanted behavior
         this.$emit('dataChange')
-      }
+      }, debounceTime)
     },
     computed: {
       customValidation () {
-        let customValidation = {'validate': this.isValid}
-        if (this.field.type === 'number' && this.field.range) {
-          customValidation.range = this.isWithinRange
+        let validate = {'validate': this.isValid}
+        if (this.isNumberField(this.field)) {
+          if (this.field.type === 'integer') {
+            validate = { ...validate, integer: this.isCompatibleWithJavaInt() }
+          } else if (this.field.type === 'long') {
+            validate = { ...validate, long: this.isCompatibleWithJavaLong() }
+          }
         }
-        return customValidation
+
+        if (this.isNumberField(this.field) && this.field.range) {
+          validate = { ...validate, range: this.isWithinRange }
+        }
+
+        return validate
+      },
+      stepSize () {
+        // Conditionally add step size, return false to omit step attribute
+        return (this.field.type === 'integer' || this.field.type === 'long') ? 1 : false
+      },
+      inputType () {
+        return this.isNumberField(this.field) ? 'number' : this.field.type
       }
     },
     methods: {
@@ -92,7 +126,19 @@
         }
 
         return true
+      },
+      isCompatibleWithJavaInt () {
+        return Number.isSafeInteger(this.value) && this.value <= MAX_JAVA_INT && this.value >= MIN_JAVA_INT
+      },
+      isCompatibleWithJavaLong () {
+        return Number.isInteger(this.value)
+      },
+      isNumberField (field) {
+        return field.type === 'integer' || field.type === 'long' || field.type === 'decimal'
       }
+    },
+    created () {
+      debounceTime = this.inputDebounceTime
     }
   }
 </script>